move insurance provider types to general crate
