+++ server/graphql/types/src/types/mod.rs
@@ -28,9 +28,6 @@ pub use self::location::*;
 pub mod master_list;
 pub use self::master_list::*;
 
-pub mod insurance_provider;
-pub use self::insurance_provider::*;
-
 pub mod invoice_query;
 pub use self::invoice_query::*;
